feat: create ColorNew component
This component shows a form to add a new color.  It requires a color
name and color value.

A new function, to save the color to state, is added to the main app
component.

The list of colors in ColorList.js is reversed so that the last color
added is displayed on top.  The colors state has no reversing
functionality on it, because newly added colors will always be added to
the end, causing any reversing to be needed after adding every new
color.

This completes part 2.