fix: config jest to transform modules
!IMPORTANT FIX!

When running tests, imports, like for Axios, may cause this error to
occur:
SyntaxError: Cannot use import statement outside a module

To fix this, Jest needs to be configured to ignore all node modules
(which is the default) except the offending modules.  The configuration
can be placed in package.json.